requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.11
aiosignal==1.3.2
aiofiles==24.1.0
async-lru==2.0.5
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.14.2
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
crypto==1.4.1
cycler==0.12.1
discord.py-self==2.0.1
fonttools==4.58.4
frozenlist==1.5.0
gevent==25.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
idna==3.10
jiter==0.10.0
kiwisolver==1.4.8
matplotlib==3.10.3
multidict==6.3.2
Naked==0.1.32
numpy==2.2.6
openai==1.80.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.30.2
pybit==5.11.0
pycrypto==2.6.1
pycryptodome==3.23.0
pydantic==2.11.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==20.7
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
setuptools==80.9.0
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket==0.2.1
websocket-client==1.8.0
websockets==15.0.1
yarl==1.19.0
zope.event==5.1
zope.interface==7.2
